Problem3.cpp:     - Removed warning. Problem8.cpp:     - Renamed printResults to print_results.     - Moved test_function_performance into a namespace. Problem11.cpp:     - Minor changes to version 3.     - Added book's version.     - Added test_function_performance.

+++ TheModernC++Challenge/src/Chapter1_MathProblems/Problem3.cpp
@@ -15,7 +15,7 @@ auto lcm(const std::vector<size_t>& v)
             return accum + possible_lcm % n;
         };
 
-        size_t sum_of_mods = std::accumulate(std::next(v.crbegin()), v.crend(), 0, accum_plus_possible_lcm_mod_n);
+        size_t sum_of_mods = std::accumulate(std::next(v.crbegin()), v.crend(), static_cast<size_t>(0), accum_plus_possible_lcm_mod_n);
 
         if (sum_of_mods == 0)
         {
